:adhesive_bandage: SessionService Interface 생성
